Fix integration tests failing in CI
Tests now check for the CI envar to enable CI features instead of using
--ci flags.

+++ integration-test/run-all.sh
@@ -1,7 +1,26 @@
 #!/usr/bin/env bash
 set -euo pipefail
+NO_INSTALL="$(printf '%s\n' "$@" | grep '^--no-install$' || true)"
 DIR="$(readlink -f "$(dirname "${BASH_SOURCE[0]}")")"
+
 cd "$DIR"
-npm uninstall --no-save @lib.cam/xslt-nailgun
-npm install
-npm run integration-test
+
+if [[ ! $NO_INSTALL ]]; then
+  ./link-local-build.sh
+  npm uninstall --no-save @lib.cam/xslt-nailgun
+  npm install
+fi
+
+if [[ ${CI:-} = true ]]; then
+  mkdir -p test-reports
+
+  set +e
+  node_modules/.bin/bats integration-tests.bats | tee test-reports/bats.tap
+  STATUS=$?
+  set -e
+
+  node_modules/.bin/tap-xunit < test-reports/bats.tap > test-reports/bats.xml
+  exit $STATUS
+else
+  node_modules/.bin/bats integration-tests.bats
+fi
